style(optionbuttons): move common css to src common
