+++ src/components/Filters/Filters.tsx
@@ -1,5 +1,6 @@
 import React, { useState } from 'react';
-import styles from './Filters.module.css';
+import styles from '../../common/commonOptionButton.module.css';
+import filterStyles from './Filters.module.css';
 
 const filterOptions = ['Animals', 'Health', 'Education', 'Environment'];
 
@@ -20,7 +21,7 @@ function Filters() {
    };
 
    return (
-      <div className={styles.filters}>
+      <div className={filterStyles.filters}>
          <button
             className={`${styles.button} ${isOpen ? styles.open : null}`}
             onClick={toggleIsOpen}>
@@ -34,7 +35,9 @@ function Filters() {
                   <li
                      key={option}
                      className={`${styles.optionItem} ${
-                        selectedFilters.includes(option) ? styles.selected : ''
+                        selectedFilters.includes(option)
+                           ? filterStyles.selected
+                           : ''
                      }`}
                      onClick={() => handleFilterClick(option)}>
                      {selectedFilters.includes(option) ? `✓ ${option}` : option}
